Move filter sidebar primitives from ui/ to collections/
Filter sidebar is specific to collection/search filtering, not a
generic UI primitive. Renamed to filter-primitives.tsx in collections/.
